+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to the versioning scheme outlined in the [README.md](RE
 
 - When a miner times out waiting for signatures, it will re-propose the same block instead of building a new block ([#5877](https://github.com/stacks-network/stacks-core/pull/5877))
 
+- Removed "fee" argument from `TransactionResult::success()` and `TransactionResult::success_with_soft_limit()` methods ([#5900](https://github.com/stacks-network/stacks-core/issues/5900))
+
 ## [3.1.0.0.7]
 
 ### Added